Split DataStacker into a separate file, start on typehints.
