+++ datacube_ows/wcs1_utils.py
@@ -14,7 +14,7 @@ from ows.util import Version
 from rasterio import MemoryFile
 
 from datacube_ows.cube_pool import cube
-from datacube_ows.data import DataStacker
+from datacube_ows.loading import DataStacker
 from datacube_ows.mv_index import MVSelectOpts
 from datacube_ows.ogc_exceptions import WCS1Exception
 from datacube_ows.ogc_utils import ConfigException
@@ -23,7 +23,7 @@ from datacube_ows.resource_limits import ResourceLimited
 from datacube_ows.wcs_utils import get_bands_from_styles
 
 
-class WCS1GetCoverageRequest():
+class WCS1GetCoverageRequest:
     version = Version(1, 0, 0)
     # pylint: disable=too-many-instance-attributes, too-many-branches, too-many-statements, too-many-locals
     def __init__(self, args):
